refactored users into user.js
