+++ frontend/src/core/comments/actions.js
@@ -7,24 +7,6 @@ import * as notification from 'core/notification';
 import * as history from 'modules/history';
 import * as teamcomments from 'modules/teamcomments';
 
-import type { UsersList } from 'core/api';
-
-export const RECEIVE: 'users/RECEIVE' = 'users/RECEIVE';
-
-
-export type ReceiveAction = {|
-    +type: typeof RECEIVE,
-    +users: Array<UsersList>,    
-|}
-
-export function receive(
-    users: Array<UsersList>,
-): ReceiveAction {
-    return {
-        type: RECEIVE,
-        users,
-    };
-}
 
 export function addComment(
     entity: number,
@@ -50,15 +32,7 @@ export function addComment(
     }
 }
 
-export function get(): Function {
-    return async dispatch => {
-        const content = await api.comment.getUsers();
-        dispatch(receive(content));
-    }
-}
-
 
 export default {
     addComment,
-    get,
 };
