+++ frontend/src/core/user/reducer.js
@@ -1,11 +1,13 @@
 /* @flow */
 
-import { UPDATE, UPDATE_SETTINGS } from './actions';
+import { RECEIVE, UPDATE, UPDATE_SETTINGS } from './actions';
 
-import type { UpdateAction, UpdateSettingsAction } from './actions';
+import type { ReceiveAction, UpdateAction, UpdateSettingsAction } from './actions';
+import type { UsersList } from 'core/api';
 
 
 type Action =
+    | ReceiveAction
     | UpdateAction
     | UpdateSettingsAction
 ;
@@ -91,6 +93,7 @@ export type UserState = {|
     +gravatarURLSmall: string,
     +gravatarURLBig: string,
     +notifications: Notifications,
+    +users: Array<UsersList>,
 |};
 
 const initial: UserState = {
@@ -114,6 +117,7 @@ const initial: UserState = {
         has_unread: false,
         notifications: [],
     },
+    users: [],
 };
 
 export default function reducer(
@@ -121,8 +125,14 @@ export default function reducer(
     action: Action,
 ): UserState {
     switch (action.type) {
+        case RECEIVE:
+            return {
+                ...state,
+                users: action.users,
+            };
         case UPDATE:
             return {
+                ...state,
                 isAuthenticated: action.data.is_authenticated,
                 isAdmin: action.data.is_admin,
                 id: action.data.id,
